@@ -2,7 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Kelas;
 use Illuminate\Http\Request;
+//tambahan
+use Illuminate\Support\Facades\DB;
+use App\Models\Siswa;
+use App\Models\User;
+use Illuminate\Support\Facades\Hash;
 
 class SiswaController extends Controller
 {
@@ -12,8 +18,12 @@ class SiswaController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
-        //
+    {      
+        $siswaJoin =  DB::table('siswa')
+                    ->join('kelas', 'siswa.kelas_id', '=', 'kelas.id')                    
+                    ->select('siswa.*', 'kelas.nama_kelas')
+                    ->get();
+        return view('siswa.index', compact('siswaJoin'));
     }
 
     /**
@@ -23,7 +33,7 @@ class SiswaController extends Controller
      */
     public function create()
     {
-        //
+        // return view('siswa.form');
     }
 
     /**
@@ -34,7 +44,19 @@ class SiswaController extends Controller
      */
     public function store(Request $request)
     {
-        //
+    
+        // $request->validate([
+        //     'nama' => 'required',
+        //     'nip' => 'required|uniqu:Siswa|min:6',
+        //     'no_telp' => 'required',
+        //     'email' => 'required|email',
+        //     'user_id' => 'required'            
+        // ]);
+        
+        // Siswa::create($request->all());
+
+        // return redirect()->route('siswa.index')
+        //     ->with('success', 'Data Siswa Baru Berhasil Disimpan');
     }
 
     /**
@@ -45,7 +67,7 @@ class SiswaController extends Controller
      */
     public function show($id)
     {
-        //
+       //
     }
 
     /**
@@ -56,7 +78,10 @@ class SiswaController extends Controller
      */
     public function edit($id)
     {
-        //
+        $row = Siswa::find($id);
+        $user = User::where('id', $row->user_id)->first(); 
+        $kelas = Kelas::all();
+        return view('profile.form_edit', compact('row','user','kelas'));
     }
 
     /**
@@ -67,8 +92,27 @@ class SiswaController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
-    {
-        //
+    {   
+        
+        $data_siswa = [
+            'nama' => $request->nama,
+            'nisn' => $request->nisn,
+            'nama_ortu' => $request->nama_ortu,
+            'no_telp' => $request->no_telp,
+            'kelas_id' => $request->kelas_id,
+        ];
+
+        $data_user = [
+            'email' => $request->email,
+            'username' => $request->username,
+            'password' => Hash::make($request->password)
+        ];
+        
+        Siswa::find($id)->update($data_siswa);
+        User::find($request->user_id)->update($data_user);
+
+        return redirect()->route('siswa.edit',$id)
+            ->with('success', 'Profile Berhasil Diubah');
     }
 
     /**
@@ -79,6 +123,23 @@ class SiswaController extends Controller
      */
     public function destroy($id)
     {
-        //
+        
+        $row = Siswa::find($id);
+        
+        Siswa::where('id', $id)->delete();
+        return redirect()->route('siswa.index')
+            ->with('success', 'Data Siswa Berhasil Dihapus');
     }
+
+    // public function SiswaPDF()
+    // {
+    //     $Siswa = Siswa::all();
+    //     $pdf = PDF::loadView('Siswa.SiswaPDF', ['Siswa' => $Siswa]);
+    //     return $pdf->download('data_Siswa.pdf');
+    // }
+
+    // public function SiswaExcel()
+    // {
+    //     return Excel::download(new SiswaExport, 'data_Siswa.xlsx');
+    // }
 }
