@@ -7,9 +7,22 @@ use Illuminate\Database\Eloquent\Model;
 
 class Guru extends Model
 {
+    /**
+     * The table associated with the model.
+     *
+     * @var string
+     */
     use HasFactory;
     //mapping ke tabel
-    protected $table = 'Guru';
+    protected $table = 'guru';
     //mapping ke kolom/fieldnya
-    protected $fillable = ['Nama','NIP','Email','Jenis Kelamin','No Telepon'];
+    protected $fillable = [
+        'nama', 'nip', 'no_telp', 'user_id'
+    ];
+    
+    public $timestamps = false;
+    public function guru()
+    {
+        return $this->hasMany(Guru::class);
+    }
 }
